move mesh generation methods to YarnMeshGenerator class
