+++ KnittingChartPreview/Assets/Scripts/YarnMeshGenerator.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 181d04a78e9b408dae2a872ba3f8c070
+timeCreated: 1711624477
